Gateways restructure and redirect logic
