+++ src/nextapp/pages/manager/gateways/detail.tsx
